addressing all comment from Lawrence
